училась работать с сервлетами, создала два сервлета, один выводит сумму зарплат преподавателей, второй - список студентов
